pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["prStudentka <bk_nvk@mail.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]
classifiers = []
repository = "https://github.com/prStudentka/python-project-52"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1.3"
django-bootstrap5 = "^23.3"
python-dotenv = "^1.0.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
django-filter = "^23.5"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
